requirements.txt
azure-cognitiveservices-speech>=1.40.0
azure-identity>=1.17.0
requests>=2.31.0
flask>=3.0.0
flask-socketio>=5.3.0
gunicorn>=22.0.0
gevent>=24.2.0
gevent-websocket>=0.10.1
